LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-anchor-download-userInitiated-expected.txt

Click me

FAIL <a download> click fires navigate event assert_equals: expected (object) Element node <a id="a" href="?1" download="">Click me</a> but got (undefined) undefined
